image_forgery_detection/requirements.txt
# Basic requirements
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.1
matplotlib>=3.4.2
opencv-python>=4.5.3
scikit-learn>=0.24.2
scikit-image>=0.18.2

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
albumentations>=1.0.3
timm>=0.4.12

# Utilities
PyYAML>=5.4.1
tqdm>=4.62.0
wandb>=0.12.0
tensorboard>=2.6.0

# Image processing
scipy>=1.7.0
imageio>=2.9.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Documentation
sphinx>=4.1.2
sphinx-rtd-theme>=0.5.2